wrapping stuff in CmdSpec.
This prepares us to support computing a CmdSpec and then running it
instead of a general IO action.

The reason we want to move away from general IO action is that
we can then allow a verbose mode that tells user what exactly
is the external command to be executed.